requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-rest-swagger==0.3.10
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
